First pass at supporting WPE releases
This de-hardcodes the WebKitGTK+ port name and its details (mailing
list, IRC channels, description text, etc.) and makes them variable
depending on the value of the new "--port=NAME" command line option.

With these changes applied, it is already possible to perform some
operations on WPE tarballs. For example, the "install-tarball" and
"send-announcement" are already working fine for WPE tarballs.