Add a new "gen-tarball-meta" subcommand
This can be used to generate checksums and GPG detached signatures
for any tarball, even if they do not have a NEWS file: just pass
"--no-news" to disable its generation. This can be used e.g. to
sign releases of the WPE backend tarballs, which do not follow the
same conventions as WebKit tarballs. Also, this adds a fallback which
extracts version numbers from the file name of the tarball instead
of grepping the autotools/CMake files in the contents of the tarball.