wsa: Fetch CVE data lazily using the JSON format
Instead of fetching data for all ever published CVEs, download the
information for each needed CVE lazily, as needed. While at it, switch
to the JSON v5 CVE data format, because the CSV dumps previously used
do not allow fetching data for individual CVEs. The CSV dumps have been
deprecated in March 2023, too:

  https://www.cve.org/Media/News/item/blog/2023/03/29/CVE-Downloads-in-JSON-5-Format

This introduces a new "cve" module with helper classes to fetch (and
cache locally) CVE JSON data on demand, and makes the "wsa" script use
the new module for the "fill" and "generate" subcommands. As the CVE
data fetches now happen on demand, the "wsa cve-data" subcommand used
to fetch a CSV dump is no longer needed.