wsa: Cache HTTP fetches in check subcommand
Introduce a helper httpcache Python module, then use it to fetch
advisory pages. The cache tracks the Last-Modified and ETag of
downloaded content, which then reuses when re-fetching data, giving
servers the chance to respond with the HTTP 304 code (not modified).

Actual content is stored in content-addressable blobs identified by
the SHA256 checksum of their contained data. While unlikely to happe
n for the HTML pages of the Apple advisories, this sets the scene for
deduplicated cached content.

This cuts typical crawl times to one third of the time, e.g. a typical
run goes from about ~15 minutes to a little below 5 minutes.